+++ app/src/main/res/layout/input_fragment.xml
@@ -1,27 +1,12 @@
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/fragment_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@drawable/hangman1"
-    tools:context="asd.org.ahorcado.fragments.InputFragment">
-
-    <GridLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="horizontal">
-
-        <ProgressBar
-            android:id="@+id/progressBar"
-            style="?android:attr/progressBarStyleHorizontal"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="left"
-            android:progressTint="@color/colorRed" />
-    </GridLayout>
+    android:background="@android:color/transparent"
+    tools:context=".InputFragment">
 
     <TableLayout
-        android:id="@+id/linearLayout1"
+        android:id="@+id/tableLayout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_gravity="left|bottom">
