+++ app/src/main/res/layout/word_fragment.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".WordFragment">
+
+    <ImageView
+        android:id="@+id/image_view"
+        android:layout_width="175dp"
+        android:layout_height="272dp"
+        android:layout_gravity="center" />
+
+</FrameLayout>
